validate email and password
