--- routes/accounts.js
@@ -1,29 +1,0 @@
-const express = require('express')
-const router = express.Router()
-
-// Routes below are prepended with /account from mounting on app.js
-router.get('/login', function (req, res) {
-  res.render('login')
-})
-
-router.post('/login', function (req, res) {
-  res.send('Attemping to send login information')
-})
-
-router.get('/register', function (req, res) {
-  res.render('register')
-})
-
-router.post('/register', function (req, res) {
-  res.send('Attempting to send register information')
-})
-
-router.get('/edit', function (req, res) {
-  res.send('Account settings')
-})
-
-router.post('/edit', function (req, res) {
-  res.send('Change account settings')
-})
-
-module.exports = router // exports this router usually to app.js
